+++ process/admin/sample/sample1_p.php
@@ -1,9 +1,8 @@
 <?php 
-include '../../conn.php';
+require '../../DatabaseConnections.php';
 
 if (!isset($_POST['method'])) {
 	echo 'method not set';
-	exit;
 }
 
 $method = $_POST['method'];
@@ -11,25 +10,43 @@ $method = $_POST['method'];
 if ($method == 'account_list') {
     $data = array();
 
-	$query = "SELECT * FROM user_accounts";
-	$stmt = $conn->prepare($query, array(PDO::ATTR_CURSOR => PDO::CURSOR_SCROLL));
-	$stmt->execute();
-	if ($stmt->rowCount() > 0) {
-		foreach($stmt->fetchALL() as $row){
+	// Connection Object
+    $conn = null;
+
+    // Connection Open
+    $connectionArr = $db->connect();
+
+    if ($connectionArr['connected'] == 1) {
+        $conn = $connectionArr['connection'];
+
+		$query = "SELECT * FROM user_accounts";
+		$stmt = $conn->prepare($query, array(PDO::ATTR_CURSOR => PDO::CURSOR_SCROLL));
+		$stmt->execute();
+		if ($stmt->rowCount() > 0) {
+			foreach($stmt->fetchALL() as $row) {
+				$data[] = array(
+					'id' => $row['id'],
+					'id_number' => $row['id_number'],
+					'username' => htmlspecialchars($row['username']),
+					'full_name' => htmlspecialchars($row['full_name']),
+					'section' => $row['section'],
+					'role' => $row['role']
+				);
+			}
+		} else {
 			$data[] = array(
-				'id' => $row['id'],
-				'id_number' => $row['id_number'],
-				'username' => htmlspecialchars($row['username']),
-				'full_name' => htmlspecialchars($row['full_name']),
-                'section' => $row['section'],
-                'role' => $row['role']
+				'message' => 'No Results Found'
 			);
 		}
-	}else{
+	} else {
 		$data[] = array(
-			'message' => 'No Results Found'
+			'message' => $connectionArr['title'] . " " . $connectionArr['message']
 		);
-	}
+    }
+
+    // Connection Close
+    $conn = null;
+
     header('Content-Type: application/json; charset=utf-8');
 	echo json_encode($data, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES);
 }
@@ -39,26 +56,44 @@ if ($method == 'search_account_list') {
 	$full_name = $_POST['full_name'];
 	$user_type = $_POST['user_type'];
 
-	$query = "SELECT * FROM user_accounts WHERE id_number LIKE '$employee_no%' 
+	// Connection Object
+    $conn = null;
+
+    // Connection Open
+    $connectionArr = $db->connect();
+
+    if ($connectionArr['connected'] == 1) {
+        $conn = $connectionArr['connection'];
+
+		$query = "SELECT * FROM user_accounts WHERE id_number LIKE '$employee_no%' 
             AND full_name LIKE '$full_name%' AND role LIKE '$user_type%'";
-	$stmt = $conn->prepare($query, array(PDO::ATTR_CURSOR => PDO::CURSOR_SCROLL));
-	$stmt->execute();
-	if ($stmt->rowCount() > 0) {
-		foreach($stmt->fetchALL() as $row){
+		$stmt = $conn->prepare($query, array(PDO::ATTR_CURSOR => PDO::CURSOR_SCROLL));
+		$stmt->execute();
+		if ($stmt->rowCount() > 0) {
+			foreach($stmt->fetchALL() as $row) {
+				$data[] = array(
+					'id' => $row['id'],
+					'id_number' => $row['id_number'],
+					'username' => htmlspecialchars($row['username']),
+					'full_name' => htmlspecialchars($row['full_name']),
+					'section' => $row['section'],
+					'role' => $row['role']
+				);
+			}
+		} else {
 			$data[] = array(
-				'id' => $row['id'],
-				'id_number' => $row['id_number'],
-				'username' => htmlspecialchars($row['username']),
-				'full_name' => htmlspecialchars($row['full_name']),
-                'section' => $row['section'],
-                'role' => $row['role']
+				'message' => 'No Results Found'
 			);
 		}
-	}else{
+	} else {
 		$data[] = array(
-			'message' => 'No Results Found'
+			'message' => $connectionArr['title'] . " " . $connectionArr['message']
 		);
-	}
+    }
+
+    // Connection Close
+    $conn = null;
+
     header('Content-Type: application/json; charset=utf-8');
 	echo json_encode($data, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES);
 }
@@ -71,24 +106,39 @@ if ($method == 'register_account') {
 	$section = trim($_POST['section']);
 	$user_type = trim($_POST['user_type']);
 
-	$check = "SELECT id FROM user_accounts WHERE username = ?";
-	$stmt = $conn->prepare($check, array(PDO::ATTR_CURSOR => PDO::CURSOR_SCROLL));
-	$params = array($username);
-	$stmt->execute($params);
-	if ($stmt->rowCount() > 0) {
-		echo 'Already Exist';
-	}else{
-		$stmt = NULL;
-		$query = "INSERT INTO user_accounts (id_number, full_name, username, password, section, role) 
-                VALUES (?, ?, ?, ?, ?, ?)";
-		$stmt = $conn->prepare($query);
-		$params = array($employee_no, $full_name, $username, $password, $section, $user_type);
-		if ($stmt->execute($params)) {
-			echo 'success';
-		}else{
-			echo 'error';
+	// Connection Object
+    $conn = null;
+
+    // Connection Open
+    $connectionArr = $db->connect();
+
+    if ($connectionArr['connected'] == 1) {
+        $conn = $connectionArr['connection'];
+
+		$check = "SELECT id FROM user_accounts WHERE username = ?";
+		$stmt = $conn->prepare($check, array(PDO::ATTR_CURSOR => PDO::CURSOR_SCROLL));
+		$params = array($username);
+		$stmt->execute($params);
+		if ($stmt->rowCount() > 0) {
+			echo 'Already Exist';
+		} else {
+			$query = "INSERT INTO user_accounts 
+					(id_number, full_name, username, password, section, role) 
+					VALUES (?, ?, ?, ?, ?, ?)";
+			$stmt = $conn->prepare($query);
+			$params = array($employee_no, $full_name, $username, $password, $section, $user_type);
+			if ($stmt->execute($params)) {
+				echo 'success';
+			} else {
+				echo 'error';
+			}
 		}
-	}
+	} else {
+        echo $connectionArr['title'] . " " . $connectionArr['message'];
+    }
+
+    // Connection Close
+    $conn = null;
 }
 
 if ($method == 'update_account') {
@@ -100,37 +150,78 @@ if ($method == 'update_account') {
 	$section = trim($_POST['section']);
 	$role = trim($_POST['role']);
 
-	$query = "SELECT id FROM user_accounts WHERE username = ?";
-	$stmt = $conn->prepare($query, array(PDO::ATTR_CURSOR => PDO::CURSOR_SCROLL));
-	$params = array($username);
-	$stmt->execute($params);
-	if ($stmt->rowCount() > 0) {
-		echo 'duplicate';
-	}else{
-		$stmt = NULL;
-		$query = "UPDATE user_accounts SET id_number = ?, username = ?, full_name = ?, 
-                password = ?, section = ?, role = ? WHERE id = ?";
-		$stmt = $conn->prepare($query);
-        $params = array($id_number, $username, $full_name, $password, $section, $role, $id);
-		if ($stmt->execute($params)) {
-			echo 'success';
-		}else{
-			echo 'error';
+	// Connection Object
+    $conn = null;
+
+    // Connection Open
+    $connectionArr = $db->connect();
+
+    if ($connectionArr['connected'] == 1) {
+        $conn = $connectionArr['connection'];
+
+		$query = "SELECT id FROM user_accounts WHERE username = ?";
+		$stmt = $conn->prepare($query, array(PDO::ATTR_CURSOR => PDO::CURSOR_SCROLL));
+		$params = array($username);
+		$stmt->execute($params);
+		if ($stmt->rowCount() > 0) {
+			echo 'duplicate';
+		} else {
+			$query = "UPDATE user_accounts SET id_number = ?, username = ?, 
+					full_name = ?, section = ?, role = ?";
+			if (!empty($password)) {
+				$query .= ", password = ?";
+			}
+			$query .= " WHERE id = ?";
+
+			$params = array();
+			
+			if (!empty($password)) {
+				$params = array($id_number, $username, $full_name, $section, $role, $password, $id);
+			} else {
+				$params = array($id_number, $username, $full_name, $section, $role, $id);
+			}
+
+			$stmt = $conn->prepare($query);
+			if ($stmt->execute($params)) {
+				echo 'success';
+			} else {
+				echo 'error';
+			}
 		}
-	}
+	} else {
+        echo $connectionArr['title'] . " " . $connectionArr['message'];
+    }
+
+    // Connection Close
+    $conn = null;
 }
 
 if ($method == 'delete_account') {
 	$id = $_POST['id'];
 
-	$query = "DELETE FROM user_accounts WHERE id = ?";
-	$stmt = $conn->prepare($query);
-    $params = array($id);
-	if ($stmt->execute($params)) {
-		echo 'success';
-	}else{
-		echo 'error';
-	}
+	// Connection Object
+    $conn = null;
+
+    // Connection Open
+    $connectionArr = $db->connect();
+
+    if ($connectionArr['connected'] == 1) {
+        $conn = $connectionArr['connection'];
+
+		$query = "DELETE FROM user_accounts WHERE id = ?";
+		$stmt = $conn->prepare($query);
+		$params = array($id);
+		if ($stmt->execute($params)) {
+			echo 'success';
+		} else {
+			echo 'error';
+		}
+	} else {
+        echo $connectionArr['title'] . " " . $connectionArr['message'];
+    }
+
+    // Connection Close
+    $conn = null;
 }
 
 if ($method == 'delete_account_selected') {
@@ -138,18 +229,31 @@ if ($method == 'delete_account_selected') {
 	$id_arr = $_POST['id_arr'];
 
 	$count = count($id_arr);
-	foreach ($id_arr as $id) {
-		$sql = "DELETE FROM user_accounts WHERE id = ?";
-		$stmt = $conn -> prepare($sql);
-		$params = array($id);
-		$stmt -> execute($params);
-		$count--;
-	}
+
+	// Connection Object
+    $conn = null;
+
+    // Connection Open
+    $connectionArr = $db->connect();
+
+    if ($connectionArr['connected'] == 1) {
+        $conn = $connectionArr['connection'];
+
+		foreach ($id_arr as $id) {
+			$sql = "DELETE FROM user_accounts WHERE id = ?";
+			$stmt = $conn -> prepare($sql);
+			$params = array($id);
+			$stmt -> execute($params);
+			$count--;
+		}
+	} else {
+        echo $connectionArr['title'] . " " . $connectionArr['message'];
+    }
+
+    // Connection Close
+    $conn = null;
 
 	if ($count == 0) {
 		echo 'success';
 	}
 }
-
-$conn = NULL;
-?>
